C4-1019: continue ingestion-to-snovault migration
